% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{hist_plot}
\alias{hist_plot}
\title{Histogram custom function}
\usage{
hist_plot(df, x_var, bins_var)
}
\arguments{
\item{df}{dataset}

\item{x_var}{x variable}

\item{bins_var}{bins (numeric)}
}
\value{
ggplot2 plot
}
\description{
Histogram custom function
}
