% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch bslib (examples)}
\usage{
launch_app(app, run = "w", options = list())
}
\arguments{
\item{app}{one of:
\itemize{
\item \code{"basic"}
\item \code{"nested_sidebar"}
\item \code{"sidebar_toggle"}
\item \code{"sidebar_cond_panel"}
}}

\item{run}{where to launch app:
\itemize{
\item \code{p} = launch in viewer pane
\item \code{b} = launch in external browser
\item \code{w} = launch in window
}}

\item{options}{arguments to pass to \code{options()}}
}
\value{
Shiny app object
}
\description{
Launch bslib (examples)
}
