% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_server}
\alias{mod_data_server}
\title{Module (data) server}
\usage{
mod_data_server(id, .dev = FALSE)
}
\arguments{
\item{id}{module id}

\item{.dev}{view reactive inputs}
}
\value{
module server function
}
\description{
Module (data) server
}
