% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input.R
\name{mod_inputs_server}
\alias{mod_inputs_server}
\title{input shiny module (server)}
\usage{
mod_inputs_server(id, .dev = FALSE)
}
\arguments{
\item{id}{shiny module id}
}
\description{
input shiny module (server)
}
