% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_numeric_inputs.R
\name{mod_numeric_inputs_server}
\alias{mod_numeric_inputs_server}
\title{input shiny module (server)}
\usage{
mod_numeric_inputs_server(id, .dev = FALSE)
}
\arguments{
\item{id}{shiny module id}

\item{.dev}{view output from \code{reactiveValuesToList()}?}
}
\description{
input shiny module (server)
}
