% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_numeric_inputs.R
\name{mod_numeric_inputs_ui}
\alias{mod_numeric_inputs_ui}
\title{input shiny module (UI)}
\usage{
mod_numeric_inputs_ui(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
input shiny module (UI)
}
