% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_server}
\alias{mod_plot_server}
\title{Module (plot) server}
\usage{
mod_plot_server(id, vals, .dev = FALSE)
}
\arguments{
\item{id}{module id}

\item{vals}{reactive values from \code{mod_data_server()}}

\item{.dev}{view reactive inputs}
}
\value{
module server function
}
\description{
Module (plot) server
}
