% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_text_inputs.R
\name{mod_text_inputs_server}
\alias{mod_text_inputs_server}
\title{text_inputs shiny module (server)}
\usage{
mod_text_inputs_server(id, .dev = FALSE)
}
\arguments{
\item{id}{shiny module id}
}
\description{
text_inputs shiny module (server)
}
