% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_text_inputs.R
\name{mod_text_inputs_ui}
\alias{mod_text_inputs_ui}
\title{text_inputs shiny module (UI)}
\usage{
mod_text_inputs_ui(id)
}
\arguments{
\item{id}{shiny module id}
}
\description{
text_inputs shiny module (UI)
}
